pyproject.toml
[tool.poetry]
name = "notebook-llm-agents"
version = "0.1.0"
description = ""
authors = ["konstantgr <konstantin.grotov@gmail.com>"]
readme = "README.md"

packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
grazie-api-gateway-client = "0.0.18"
rpdb = "^0.1.6"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.1"
nbformat = "^5.9.2"
omegaconf = "^2.3.0"
nbconvert = "^7.15.0"

[[tool.poetry.source]]
name = "space-grazie-ml"
url = "https://packages.jetbrains.team/pypi/p/grazi/grazie-ml/simple"


[tool.poetry.group.server.dependencies]
pydantic = "^2.6.1"
fastapi = "^0.109.2"
requests = "^2.31.0"
uvicorn = "^0.27.0.post1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
